requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-filter==2.3.0
django-mysql==3.7.0
django-pandas==0.6.2
django-redis==4.12.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
numpy==1.19.0
openpyxl==3.0.3
pandas==1.0.5
PyJWT==1.7.1
PyMySQL==0.9.3
redis==3.5.3
rest-pandas==1.1.0
six==1.15.0
urllib3==1.25.9
uWSGI==2.0.18
